avoid another cyclic import
