+++ nicegui/globals.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import asyncio
-import inspect
 import logging
 import os
 from contextlib import contextmanager
@@ -11,9 +10,6 @@ from typing import TYPE_CHECKING, Any, Awaitable, Callable, Dict, Iterator, List
 
 from socketio import AsyncServer
 from uvicorn import Server
-
-from . import background_tasks
-from .helpers import is_coroutine_function
 
 if TYPE_CHECKING:
     from .air import Air
@@ -118,11 +114,3 @@ def socket_id(id_: str) -> Iterator[None]:
     _socket_id = id_
     yield
     _socket_id = None
-
-
-def handle_exception(exception: Exception) -> None:
-    """Handle an exception by invoking all registered exception handlers."""
-    for handler in exception_handlers:
-        result = handler() if not inspect.signature(handler).parameters else handler(exception)
-        if is_coroutine_function(handler):
-            background_tasks.create(result)
